pipelines/pipeline_xero_to_postgres/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.7.3
    # via pipeline-xero-to-postgres (setup.py)
async-timeout==3.0.1
    # via aiohttp
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
backoff==1.10.0
    # via pipelinewise-singer-python
chardet==3.0.4
    # via aiohttp
ciso8601==2.3.2
    # via pipelinewise-singer-python
idna==3.10
    # via yarl
inflection==0.3.1
    # via pipelinewise-target-postgres
joblib==1.2.0
    # via pipelinewise-target-postgres
jsonschema==3.2.0
    # via pipelinewise-singer-python
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
pipelinewise-singer-python==1.3.0
    # via
    #   pipeline-xero-to-postgres (setup.py)
    #   pipelinewise-target-postgres
pipelinewise-target-postgres==2.1.2
    # via pipeline-xero-to-postgres (setup.py)
propcache==0.3.0
    # via yarl
psycopg2-binary==2.9.5
    # via pipelinewise-target-postgres
pyrsistent==0.20.0
    # via jsonschema
python-dateutil==2.9.0.post0
    # via pipelinewise-singer-python
python-dotenv==0.21.0
    # via pipeline-xero-to-postgres (setup.py)
pytz==2020.5
    # via pipelinewise-singer-python
pyyaml==6.0
    # via pipeline-xero-to-postgres (setup.py)
ruff==0.11.0
    # via pipeline-xero-to-postgres (setup.py)
simplejson==3.17.2
    # via pipelinewise-singer-python
six==1.17.0
    # via
    #   jsonschema
    #   python-dateutil
tenacity==9.0.0
    # via pipeline-xero-to-postgres (setup.py)
typing-extensions==4.12.2
    # via
    #   aiohttp
    #   multidict
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
